JAMES-2218 Extract messageId handling in cucumber tests
